% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.pct.classements.grille.R
\name{geom.pct.classements.grille}
\alias{geom.pct.classements.grille}
\title{Create geom bar of the shares of each classement letter in the AFTT grille}
\usage{
geom.pct.classements.grille(grille = share.m.B2toNC(), clr = "deeppink1")
}
\arguments{
\item{grille}{Numeric vector. Cumulative shares of each classement in the AFTT grille.}

\item{clr}{Character. Color for the bars representing the shares of each classement letter in the AFTT grille. Default: "deeppink1".}
}
\value{
A \code{ggplot2} layer (geom_errorbar) representing the shares of each classement letter in the AFTT grille as horizontal bars, to be added to a plot of the distribution of active players by classement letter.
}
\description{
Create geom bar of the shares of each classement letter in the AFTT grille
}
\examples{
library(ggplot2)
ggplot()+geom.pct.classements.grille()

}
