% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.club.player.list.m.R
\name{get.club.player.list.m}
\alias{get.club.player.list.m}
\title{Retrieve players messieurs list (ranked) for a given club}
\usage{
get.club.player.list.m(indice)
}
\arguments{
\item{indice}{Character string. Club indice (e.g., \code{"N051"}).}
}
\value{
A data frame containing the players list of the club
as of the AFTT website at the moment of retrieval, sorted by ranking.
Column names follow the structure of the JSON returned by the site but player names are removed and licence number is used as unique id. Licence number is extracted from the action field (button) of the html
\describe{
\item{position}{ranking}
\item{position_bis}{ranking without inactive or 'inactive'}
\item{classement}{classement at start of season}
\item{club}{indice club (i.e. same as function argument)}
\item{match}{number of matches played so far in season}
\item{points}{points so far in season}
\item{licence}{licence number extracted from the html action button}
}
}
\description{
Submits a POST request to the AFTT club ranking page and extracts the
\emph{Messieurs} ranking. The function isolates the JSON array defining
the rows for the selected club and returns it as a data frame.
}
\note{
This function scrapes content from the AFTT website
\url{https://data.aftt.be/ranking/clubs.php} for a single club.
Only the \emph{Messieurs} category is extracted.
Uses \pkg{httr}, \pkg{rvest}, \pkg{stringr}, and
\pkg{jsonlite}.
}
\examples{
\dontrun{
playersN051 <- get.club.player.list.m("N051")
head(playersN051)
}

}
