% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.summary.R
\name{players.summary}
\alias{players.summary}
\title{Summary (wrapper) of player counts and tables}
\usage{
players.summary(players = players_m)
}
\arguments{
\item{players}{A players data.frame. Defaults to \code{players_m}, but you may
supply \code{players_f} or any archived snapshot.}
}
\value{
A list containing counts and tables for the supplied dataset.
}
\description{
Summary (wrapper) of player counts and tables
}
