% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank2classement.R
\name{rank2classement}
\alias{rank2classement}
\title{Estimates 'Messieurs' classement (B2 to NC) from ranking and number of actifs using a grid of average cumulated share of each classement as of share.m.B2toNC() by default}
\usage{
rank2classement(x, N_actifs, sh_vector = share.m.B2toNC())
}
\arguments{
\item{x}{rank of player}

\item{N_actifs}{number of active players}

\item{sh_vector}{cumulated share of each classement, defaulted to share.m.B2toNC()}
}
\value{
a character with the estimated classement, e.g. "D6"
}
\description{
Estimates 'Messieurs' classement (B2 to NC) from ranking and number of actifs using a grid of average cumulated share of each classement as of share.m.B2toNC() by default
}
\examples{
rank2classement(9350,15780)
rank2classement(9350,16780)

}
